Move Extern types from submodule to lib in FFI. For some reason, if these types are in a submodule, even if they are publically used, the functions inside the FFI are not found in Android. Works for iOS though. To be investigated later....
